consumer/requirements.txt
#requirements
kafka-python
psycopg2-binary
requests